reorganize golbat adapters
