+++ src/lib/server/queryMapObjects/queryPokemon.ts
@@ -2,8 +2,8 @@ import { shouldDisplayPokemon } from "@/lib/features/filterLogic/pokemon";
 import type { FilterPokemon } from "@/lib/features/filters/filters";
 import type { Bounds } from "@/lib/mapObjects/mapBounds";
 import { MapObjectType, type MinMapObject } from "@/lib/mapObjects/mapObjectTypes";
-import { getMultiplePokemon, getSinglePokemon } from "@/lib/server/api/golbatApi";
-import { grpcScanPokemon, scanViaGrpcOrHttp } from "@/lib/server/api/golbatGrpc";
+import { getMultiplePokemon, getSinglePokemon } from "@/lib/server/api/golbat/http";
+import { grpcScanPokemon, scanViaGrpcOrHttp } from "@/lib/server/api/golbat/grpc";
 import { requestLimits } from "@/lib/server/api/rateLimit";
 import {
 	MapObjectQuery,
@@ -13,7 +13,7 @@ import type {
 	GolbatPokemonQuery,
 	GolbatPokemonSpecies,
 	PokemonScanBody
-} from "@/lib/server/queryMapObjects/queries";
+} from "@/lib/server/api/golbat/types";
 import { getMasterPokemon } from "@/lib/services/masterfile";
 import type { FeaturePermissionContext, PermittedPolygon } from "@/lib/services/user/checkPerm";
 import type { PokemonData, PvpStats } from "@/lib/types/mapObjectData/pokemon";
